+++ src/lungo_cli/resources/web/src/lib/constants.ts
@@ -1,6 +1,5 @@
 import { env } from "$env/dynamic/public"
 
-export const HOSTNAME = env.PUBLIC_HOSTNAME
-export const HTTPS_PORT = env.PUBLIC_HTTPS_PORT
+export const BASE_URL = env.PUBLIC_BASE_URL
 export const SITE_SUBTITLE = env.PUBLIC_SITE_SUBTITLE
 export const SITE_TITLE = env.PUBLIC_SITE_TITLE